refactor(ios): split background location concerns
