+++ packages/react-native-nitro-geolocation/ios/NitroBackgroundLocation.swift
@@ -4,40 +4,6 @@ import Foundation
 import NitroModules
 import UIKit
 
-private final class NitroBackgroundLocationDelegate: NSObject, CLLocationManagerDelegate {
-    weak var owner: NitroBackgroundLocation?
-
-    init(owner: NitroBackgroundLocation) {
-        self.owner = owner
-        super.init()
-    }
-
-    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
-        owner?.handleLocations(locations)
-        owner?.applyDeferredUpdatesIfNeeded(manager)
-    }
-
-    func locationManager(_ manager: CLLocationManager, didFailWithError error: Error) {
-        owner?.handleError(error)
-    }
-
-    func locationManager(_ manager: CLLocationManager, didEnterRegion region: CLRegion) {
-        owner?.handleRegion(region, transition: .enter)
-    }
-
-    func locationManager(_ manager: CLLocationManager, didExitRegion region: CLRegion) {
-        owner?.handleRegion(region, transition: .exit)
-    }
-
-    func locationManagerDidChangeAuthorization(_ manager: CLLocationManager) {
-        owner?.handleAuthorizationChange()
-    }
-
-    func locationManager(_ manager: CLLocationManager, didChangeAuthorization status: CLAuthorizationStatus) {
-        owner?.handleAuthorizationChange()
-    }
-}
-
 class NitroBackgroundLocation: HybridNitroBackgroundLocationSpec {
     private let defaults = UserDefaults.standard
     private let locationsKey = "nitro.background.locations"
@@ -58,6 +24,7 @@ class NitroBackgroundLocation: HybridNitroBackgroundLocationSpec {
     private let motionManager = CMMotionActivityManager()
     private let motionQueue = OperationQueue()
     private let syncQueue = DispatchQueue(label: "nitro.background.sync")
+    private let httpSync = IOSBackgroundHttpSync()
     private var permissionSemaphore: DispatchSemaphore?
     private var lastSyncAt: TimeInterval = 0
 
@@ -348,7 +315,7 @@ class NitroBackgroundLocation: HybridNitroBackgroundLocationSpec {
             guard self.permissionResult().background == .granted else {
                 throw RuntimeError.error(withMessage: "Background location permission is required to register geofences")
             }
-            let sanitized = regions.map(self.sanitizedGeofence)
+            let sanitized = regions.map(sanitizedGeofence)
             self.runOnMainSync {
                 for region in sanitized {
                     let circular = CLCircularRegion(
@@ -394,7 +361,7 @@ class NitroBackgroundLocation: HybridNitroBackgroundLocationSpec {
 
     func getRegisteredGeofences() throws -> Promise<[GeofenceRegion]> {
         return Promise.async {
-            return self.geofences.map(self.bridgeSafeGeofence)
+            return self.geofences.map(bridgeSafeGeofence)
         }
     }
 
@@ -419,7 +386,7 @@ class NitroBackgroundLocation: HybridNitroBackgroundLocationSpec {
         }
     }
 
-    fileprivate func handleLocations(_ locations: [CLLocation]) {
+    func handleLocations(_ locations: [CLLocation]) {
         for location in locations {
             let id = UUID().uuidString
             let coords = GeolocationCoordinates(
@@ -488,7 +455,7 @@ class NitroBackgroundLocation: HybridNitroBackgroundLocationSpec {
         }
     }
 
-    fileprivate func handleRegion(_ region: CLRegion, transition: GeofenceTransition) {
+    func handleRegion(_ region: CLRegion, transition: GeofenceTransition) {
         guard let geofence = geofences.first(where: { $0.identifier == region.identifier }) else {
             return
         }
@@ -564,11 +531,11 @@ class NitroBackgroundLocation: HybridNitroBackgroundLocationSpec {
         applyActivityAwareTracking(detected)
     }
 
-    fileprivate func handleAuthorizationChange() {
+    func handleAuthorizationChange() {
         permissionSemaphore?.signal()
     }
 
-    fileprivate func handleError(_ error: Error) {
+    func handleError(_ error: Error) {
         let locationError = LocationError(code: -1, message: error.localizedDescription)
         errorListeners.values.forEach { $0(locationError) }
     }
@@ -606,7 +573,7 @@ class NitroBackgroundLocation: HybridNitroBackgroundLocationSpec {
             }
             manager.desiredAccuracy = kCLLocationAccuracyBest
             manager.distanceFilter = options.distanceFilter ?? kCLDistanceFilterNone
-            manager.activityType = self.mapActivityType(options.ios?.activityType)
+            manager.activityType = mapActivityType(options.ios?.activityType)
             if options.trackingMode == .significantchanges ||
                 options.ios?.useSignificantChanges == true {
                 manager.startMonitoringSignificantLocationChanges()
@@ -616,7 +583,7 @@ class NitroBackgroundLocation: HybridNitroBackgroundLocationSpec {
         }
     }
 
-    fileprivate func applyDeferredUpdatesIfNeeded(_ manager: CLLocationManager) {
+    func applyDeferredUpdatesIfNeeded(_ manager: CLLocationManager) {
         guard
             let options,
             let distance = options.ios?.deferredUpdatesDistance,
@@ -681,7 +648,7 @@ class NitroBackgroundLocation: HybridNitroBackgroundLocationSpec {
         geofences = (defaults.array(forKey: geofencesKey) as? [[String: Any]] ?? [])
             .compactMap(makeGeofenceRegion)
         storedEvents = (defaults.array(forKey: eventsKey) as? [[String: Any]] ?? [])
-            .compactMap(makeStoredEvent)
+            .compactMap { makeStoredEvent($0, storedLocations: storedLocations) }
         options = defaults.dictionary(forKey: optionsKey).flatMap(makeBackgroundOptions)
     }
 
@@ -741,57 +708,6 @@ class NitroBackgroundLocation: HybridNitroBackgroundLocationSpec {
         defaults.set(backgroundOptionsDictionary(options), forKey: optionsKey)
     }
 
-    private func upload(
-        locations: [StoredBackgroundLocation],
-        sync: BackgroundHttpSyncOptions
-    ) -> BackgroundHttpSyncResult {
-        guard let url = URL(string: sync.url) else {
-            return BackgroundHttpSyncResult(
-                success: false,
-                statusCode: nil,
-                syncedLocationIds: [],
-                failedLocationIds: locations.map(\.id),
-                error: "Invalid sync URL"
-            )
-        }
-        var request = URLRequest(url: url)
-        request.httpMethod = sync.method?.stringValue ?? "POST"
-        request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-        sync.headers?.forEach { request.setValue($0.value, forHTTPHeaderField: $0.key) }
-        let body: Any
-        if sync.batch == false, let location = locations.first {
-            if var template = sync.bodyTemplate.map(bodyTemplateDictionary) {
-                template["location"] = storedLocationDictionary(location)
-                body = template
-            } else {
-                body = storedLocationDictionary(location)
-            }
-        } else {
-            var template = sync.bodyTemplate.map(bodyTemplateDictionary) ?? [:]
-            template["locations"] = locations.map(storedLocationDictionary)
-            body = template
-        }
-        request.httpBody = try? JSONSerialization.data(withJSONObject: body)
-
-        let semaphore = DispatchSemaphore(value: 0)
-        var statusCode: Int?
-        var requestError: String?
-        URLSession.shared.dataTask(with: request) { _, response, error in
-            statusCode = (response as? HTTPURLResponse)?.statusCode
-            requestError = error?.localizedDescription
-            semaphore.signal()
-        }.resume()
-        semaphore.wait()
-
-        let success = statusCode.map { (200..<300).contains($0) } ?? false
-        return BackgroundHttpSyncResult(
-            success: success,
-            statusCode: statusCode.map(Double.init),
-            syncedLocationIds: success ? locations.map(\.id) : [],
-            failedLocationIds: success ? [] : locations.map(\.id),
-            error: success ? nil : (requestError ?? "HTTP sync failed")
-        )
-    }
     private func scheduleSyncIfNeeded() {
         guard let sync = options?.sync else { return }
         let threshold = positiveFiniteInt(sync.syncThreshold, defaultValue: 1)
@@ -858,7 +774,7 @@ class NitroBackgroundLocation: HybridNitroBackgroundLocationSpec {
                 error: nil
             )
         }
-        let result = uploadWithRetry(locations: Array(unsynced), sync: sync)
+        let result = httpSync.uploadWithRetry(locations: Array(unsynced), sync: sync)
         if !result.success && result.syncedLocationIds.isEmpty {
             return result
         }
@@ -887,663 +803,6 @@ class NitroBackgroundLocation: HybridNitroBackgroundLocationSpec {
         }
         persistStore()
         return result
-    }
-
-    private func uploadWithRetry(
-        locations: [StoredBackgroundLocation],
-        sync: BackgroundHttpSyncOptions
-    ) -> BackgroundHttpSyncResult {
-        if sync.batch == false {
-            return uploadSingleLocationsWithRetry(locations: locations, sync: sync)
-        }
-        let maxAttempts = sync.retry == true
-            ? Int(sync.maxRetries ?? 3) + 1
-            : 1
-        var lastResult: BackgroundHttpSyncResult?
-        for attempt in 0..<max(maxAttempts, 1) {
-            let result = upload(locations: locations, sync: sync)
-            if result.success {
-                return result
-            }
-            lastResult = result
-            if attempt < maxAttempts - 1 {
-                Thread.sleep(forTimeInterval: 1)
-            }
-        }
-        return lastResult ?? BackgroundHttpSyncResult(
-            success: false,
-            statusCode: nil,
-            syncedLocationIds: [],
-            failedLocationIds: locations.map(\.id),
-            error: "HTTP sync failed"
-        )
-    }
-
-    private func uploadSingleLocationsWithRetry(
-        locations: [StoredBackgroundLocation],
-        sync: BackgroundHttpSyncOptions
-    ) -> BackgroundHttpSyncResult {
-        let maxAttempts = sync.retry == true
-            ? Int(sync.maxRetries ?? 3) + 1
-            : 1
-        var synced: [String] = []
-        var failed: [String] = []
-        var lastStatusCode: Double?
-        var lastError: String?
-
-        for location in locations {
-            var didSync = false
-            for attempt in 0..<max(maxAttempts, 1) {
-                let result = upload(locations: [location], sync: sync)
-                lastStatusCode = result.statusCode
-                lastError = result.error
-                if result.success {
-                    synced.append(location.id)
-                    didSync = true
-                    break
-                }
-                if attempt < maxAttempts - 1 {
-                    Thread.sleep(forTimeInterval: 1)
-                }
-            }
-            if !didSync {
-                failed.append(location.id)
-            }
-        }
-
-        return BackgroundHttpSyncResult(
-            success: failed.isEmpty,
-            statusCode: lastStatusCode,
-            syncedLocationIds: synced,
-            failedLocationIds: failed,
-            error: failed.isEmpty ? nil : (lastError ?? "HTTP sync failed")
-        )
-    }
-
-    private func makeStoredLocation(_ dictionary: [String: Any]) -> StoredBackgroundLocation? {
-        guard
-            let id = dictionary["id"] as? String,
-            let coordsDictionary = dictionary["coords"] as? [String: Any],
-            let coords = makeCoordinates(coordsDictionary),
-            let sourceValue = dictionary["source"] as? String,
-            let source = BackgroundLocationSource(fromString: sourceValue),
-            let recordedAt = dictionary["recordedAt"] as? Double,
-            let createdAt = dictionary["createdAt"] as? Double,
-            let timestamp = dictionary["timestamp"] as? Double
-        else {
-            return nil
-        }
-        return StoredBackgroundLocation(
-            id: id,
-            deliveredToJS: dictionary["deliveredToJS"] as? Bool ?? false,
-            synced: dictionary["synced"] as? Bool ?? false,
-            createdAt: createdAt,
-            source: source,
-            isFromBackground: dictionary["isFromBackground"] as? Bool ?? true,
-            provider: (dictionary["provider"] as? String).flatMap(LocationProviderUsed.init(fromString:)),
-            mocked: dictionary["mocked"] as? Bool,
-            recordedAt: recordedAt,
-            activity: nil,
-            battery: nil,
-            coords: coords,
-            timestamp: timestamp
-        )
-    }
-
-    private func makeStoredEvent(_ dictionary: [String: Any]) -> StoredBackgroundEventEnvelope? {
-        guard
-            let id = dictionary["id"] as? String,
-            let typeValue = dictionary["type"] as? String,
-            let type = BackgroundEventType(fromString: typeValue),
-            let timestamp = dictionary["timestamp"] as? Double,
-            let createdAt = dictionary["createdAt"] as? Double
-        else {
-            return nil
-        }
-        let persistedLocation = (dictionary["location"] as? [String: Any])
-            .flatMap(makeBackgroundLocation)
-        let referencedLocation = (dictionary["locationId"] as? String)
-            .flatMap { locationId in storedLocations.first { $0.id == locationId } }
-            .map(backgroundLocation)
-        let location = persistedLocation ?? referencedLocation
-        let geofence = (dictionary["geofence"] as? [String: Any]).flatMap(makeGeofenceEvent)
-        let activity = (dictionary["activity"] as? [String: Any]).flatMap(makeActivity)
-        let event = BackgroundEventEnvelope(
-            location: location,
-            geofence: geofence,
-            activity: activity,
-            providerStatus: nil,
-            result: (dictionary["result"] as? [String: Any]).flatMap(makeHttpSyncResult),
-            error: nil,
-            id: id,
-            type: type,
-            timestamp: timestamp,
-            deliveredToJS: dictionary["deliveredToJS"] as? Bool ?? false
-        )
-        return StoredBackgroundEventEnvelope(
-            event: event,
-            createdAt: createdAt,
-            id: id,
-            type: type,
-            timestamp: timestamp,
-            deliveredToJS: event.deliveredToJS
-        )
-    }
-
-    private func makeBackgroundLocation(_ dictionary: [String: Any]) -> BackgroundLocation? {
-        guard
-            let id = dictionary["id"] as? String,
-            let coordsDictionary = dictionary["coords"] as? [String: Any],
-            let coords = makeCoordinates(coordsDictionary),
-            let sourceValue = dictionary["source"] as? String,
-            let source = BackgroundLocationSource(fromString: sourceValue),
-            let recordedAt = dictionary["recordedAt"] as? Double,
-            let timestamp = dictionary["timestamp"] as? Double
-        else {
-            return nil
-        }
-        return BackgroundLocation(
-            id: id,
-            source: source,
-            isFromBackground: dictionary["isFromBackground"] as? Bool ?? true,
-            provider: (dictionary["provider"] as? String).flatMap(LocationProviderUsed.init(fromString:)),
-            mocked: dictionary["mocked"] as? Bool,
-            recordedAt: recordedAt,
-            activity: nil,
-            battery: nil,
-            coords: coords,
-            timestamp: timestamp
-        )
-    }
-
-    private func makeActivity(_ dictionary: [String: Any]) -> DetectedActivity? {
-        guard
-            let typeValue = dictionary["type"] as? String,
-            let type = DetectedActivityType(fromString: typeValue),
-            let confidence = dictionary["confidence"] as? Double,
-            let timestamp = dictionary["timestamp"] as? Double
-        else {
-            return nil
-        }
-        return DetectedActivity(type: type, confidence: confidence, timestamp: timestamp)
-    }
-
-    private func makeCoordinates(_ dictionary: [String: Any]) -> GeolocationCoordinates? {
-        guard
-            let latitude = dictionary["latitude"] as? Double,
-            let longitude = dictionary["longitude"] as? Double,
-            let accuracy = dictionary["accuracy"] as? Double
-        else {
-            return nil
-        }
-        return GeolocationCoordinates(
-            latitude: latitude,
-            longitude: longitude,
-            altitude: nullableDouble(dictionary["altitude"]),
-            accuracy: accuracy,
-            altitudeAccuracy: nullableDouble(dictionary["altitudeAccuracy"]),
-            heading: nullableDouble(dictionary["heading"]),
-            speed: nullableDouble(dictionary["speed"])
-        )
-    }
-
-    private func makeGeofenceRegion(_ dictionary: [String: Any]) -> GeofenceRegion? {
-        guard
-            let identifier = dictionary["identifier"] as? String,
-            let latitude = dictionary["latitude"] as? Double,
-            let longitude = dictionary["longitude"] as? Double,
-            let radius = dictionary["radius"] as? Double
-        else {
-            return nil
-        }
-        return GeofenceRegion(
-            identifier: identifier,
-            latitude: latitude,
-            longitude: longitude,
-            radius: radius,
-            notifyOnEntry: dictionary["notifyOnEntry"] as? Bool,
-            notifyOnExit: dictionary["notifyOnExit"] as? Bool,
-            notifyOnDwell: dictionary["notifyOnDwell"] as? Bool,
-            loiteringDelay: dictionary["loiteringDelay"] as? Double,
-            expirationDuration: dictionary["expirationDuration"] as? Double,
-            metadata: (dictionary["metadata"] as? [String: Any]).map(makeMetadata)
-        )
-    }
-
-    private func sanitizedGeofence(_ region: GeofenceRegion) -> GeofenceRegion {
-        return GeofenceRegion(
-            identifier: region.identifier,
-            latitude: region.latitude,
-            longitude: region.longitude,
-            radius: region.radius,
-            notifyOnEntry: region.notifyOnEntry,
-            notifyOnExit: region.notifyOnExit,
-            notifyOnDwell: region.notifyOnDwell,
-            loiteringDelay: region.loiteringDelay,
-            expirationDuration: region.expirationDuration,
-            metadata: region.metadata
-        )
-    }
-
-    private func bridgeSafeGeofence(_ region: GeofenceRegion) -> GeofenceRegion {
-        return GeofenceRegion(
-            identifier: region.identifier,
-            latitude: region.latitude,
-            longitude: region.longitude,
-            radius: region.radius,
-            notifyOnEntry: region.notifyOnEntry,
-            notifyOnExit: region.notifyOnExit,
-            notifyOnDwell: region.notifyOnDwell,
-            loiteringDelay: region.loiteringDelay,
-            expirationDuration: region.expirationDuration,
-            metadata: region.metadata
-        )
-    }
-
-    private func makeGeofenceEvent(_ dictionary: [String: Any]) -> GeofenceEvent? {
-        guard
-            let regionDictionary = dictionary["region"] as? [String: Any],
-            let region = makeGeofenceRegion(regionDictionary),
-            let transitionValue = dictionary["transition"] as? String,
-            let transition = GeofenceTransition(fromString: transitionValue),
-            let timestamp = dictionary["timestamp"] as? Double
-        else {
-            return nil
-        }
-        return GeofenceEvent(
-            region: region,
-            transition: transition,
-            location: nil,
-            timestamp: timestamp
-        )
-    }
-
-    private func storedLocationDictionary(_ location: StoredBackgroundLocation) -> [String: Any] {
-        var dictionary: [String: Any] = [
-            "id": location.id,
-            "deliveredToJS": location.deliveredToJS,
-            "synced": location.synced,
-            "createdAt": location.createdAt,
-            "source": location.source.stringValue,
-            "isFromBackground": location.isFromBackground,
-            "recordedAt": location.recordedAt,
-            "coords": coordinatesDictionary(location.coords),
-            "timestamp": location.timestamp
-        ]
-        if let provider = location.provider {
-            dictionary["provider"] = provider.stringValue
-        }
-        if let mocked = location.mocked {
-            dictionary["mocked"] = mocked
-        }
-        return dictionary
-    }
-
-    private func storedEventDictionary(_ event: StoredBackgroundEventEnvelope) -> [String: Any] {
-        var dictionary: [String: Any] = [
-            "id": event.id,
-            "type": event.type.stringValue,
-            "timestamp": event.timestamp,
-            "deliveredToJS": event.deliveredToJS,
-            "createdAt": event.createdAt
-        ]
-        if let locationId = event.event.location?.id {
-            dictionary["locationId"] = locationId
-        }
-        if let location = event.event.location {
-            dictionary["location"] = backgroundLocationDictionary(location)
-        }
-        if let geofence = event.event.geofence {
-            dictionary["geofence"] = geofenceDictionary(geofence)
-        }
-        if let activity = event.event.activity {
-            dictionary["activity"] = activityDictionary(activity)
-        }
-        if let result = event.event.result {
-            dictionary["result"] = httpSyncResultDictionary(result)
-        }
-        return dictionary
-    }
-
-    private func backgroundLocationDictionary(_ location: BackgroundLocation) -> [String: Any]? {
-        guard let id = location.id else {
-            return nil
-        }
-        var dictionary: [String: Any] = [
-            "id": id,
-            "source": location.source.stringValue,
-            "isFromBackground": location.isFromBackground,
-            "recordedAt": location.recordedAt,
-            "coords": coordinatesDictionary(location.coords),
-            "timestamp": location.timestamp
-        ]
-        if let provider = location.provider {
-            dictionary["provider"] = provider.stringValue
-        }
-        if let mocked = location.mocked {
-            dictionary["mocked"] = mocked
-        }
-        return dictionary
-    }
-
-    private func coordinatesDictionary(_ coords: GeolocationCoordinates) -> [String: Any] {
-        var dictionary: [String: Any] = [
-            "latitude": coords.latitude,
-            "longitude": coords.longitude,
-            "accuracy": coords.accuracy
-        ]
-        dictionary["altitude"] = variantDouble(coords.altitude)
-        dictionary["altitudeAccuracy"] = variantDouble(coords.altitudeAccuracy)
-        dictionary["heading"] = variantDouble(coords.heading)
-        dictionary["speed"] = variantDouble(coords.speed)
-        return dictionary
-    }
-
-    private func geofenceDictionary(_ region: GeofenceRegion) -> [String: Any] {
-        var dictionary: [String: Any] = [
-            "identifier": region.identifier,
-            "latitude": region.latitude,
-            "longitude": region.longitude,
-            "radius": region.radius
-        ]
-        dictionary["notifyOnEntry"] = region.notifyOnEntry
-        dictionary["notifyOnExit"] = region.notifyOnExit
-        dictionary["notifyOnDwell"] = region.notifyOnDwell
-        dictionary["loiteringDelay"] = region.loiteringDelay
-        dictionary["expirationDuration"] = region.expirationDuration
-        dictionary["metadata"] = region.metadata.map(metadataDictionary)
-        return dictionary
-    }
-
-    private func geofenceDictionary(_ event: GeofenceEvent) -> [String: Any] {
-        return [
-            "region": geofenceDictionary(event.region),
-            "transition": event.transition.stringValue,
-            "timestamp": event.timestamp
-        ]
-    }
-
-    private func activityDictionary(_ activity: DetectedActivity) -> [String: Any] {
-        return [
-            "type": activity.type.stringValue,
-            "confidence": activity.confidence,
-            "timestamp": activity.timestamp
-        ]
-    }
-
-    private func makeHttpSyncResult(_ dictionary: [String: Any]) -> BackgroundHttpSyncResult? {
-        guard let success = dictionary["success"] as? Bool else {
-            return nil
-        }
-        return BackgroundHttpSyncResult(
-            success: success,
-            statusCode: dictionary["statusCode"] as? Double,
-            syncedLocationIds: dictionary["syncedLocationIds"] as? [String] ?? [],
-            failedLocationIds: dictionary["failedLocationIds"] as? [String] ?? [],
-            error: dictionary["error"] as? String
-        )
-    }
-
-    private func httpSyncResultDictionary(_ result: BackgroundHttpSyncResult) -> [String: Any] {
-        var dictionary: [String: Any] = [
-            "success": result.success,
-            "syncedLocationIds": result.syncedLocationIds,
-            "failedLocationIds": result.failedLocationIds
-        ]
-        dictionary["statusCode"] = result.statusCode
-        dictionary["error"] = result.error
-        return dictionary
-    }
-
-    private func backgroundOptionsDictionary(_ options: BackgroundLocationOptions) -> [String: Any] {
-        var dictionary: [String: Any] = [:]
-        dictionary["trackingMode"] = options.trackingMode?.stringValue
-        dictionary["interval"] = options.interval
-        dictionary["fastestInterval"] = options.fastestInterval
-        dictionary["distanceFilter"] = options.distanceFilter
-        dictionary["maxUpdateDelay"] = options.maxUpdateDelay
-        dictionary["waitForAccurateLocation"] = options.waitForAccurateLocation
-        dictionary["persist"] = options.persist
-        dictionary["maxStoredLocations"] = options.maxStoredLocations
-        dictionary["maxStoredEvents"] = options.maxStoredEvents
-        dictionary["stopOnTerminate"] = options.stopOnTerminate
-        dictionary["startOnBoot"] = options.startOnBoot
-        dictionary["ios"] = options.ios.map(iosOptionsDictionary)
-        dictionary["activityRecognition"] = options.activityRecognition.map(activityOptionsDictionary)
-        dictionary["sync"] = options.sync.map(httpSyncOptionsDictionary)
-        return dictionary
-    }
-
-    private func makeBackgroundOptions(_ dictionary: [String: Any]) -> BackgroundLocationOptions {
-        return BackgroundLocationOptions(
-            trackingMode: (dictionary["trackingMode"] as? String).flatMap(BackgroundTrackingMode.init(fromString:)),
-            accuracy: nil,
-            granularity: nil,
-            interval: dictionary["interval"] as? Double,
-            fastestInterval: dictionary["fastestInterval"] as? Double,
-            distanceFilter: dictionary["distanceFilter"] as? Double,
-            maxUpdateDelay: dictionary["maxUpdateDelay"] as? Double,
-            waitForAccurateLocation: dictionary["waitForAccurateLocation"] as? Bool,
-            persist: dictionary["persist"] as? Bool,
-            maxStoredLocations: dictionary["maxStoredLocations"] as? Double,
-            maxStoredEvents: dictionary["maxStoredEvents"] as? Double,
-            stopOnTerminate: dictionary["stopOnTerminate"] as? Bool,
-            startOnBoot: dictionary["startOnBoot"] as? Bool,
-            android: nil,
-            ios: (dictionary["ios"] as? [String: Any]).map(makeIOSOptions),
-            geofencing: nil,
-            activityRecognition: (dictionary["activityRecognition"] as? [String: Any]).map(makeActivityOptions),
-            sync: (dictionary["sync"] as? [String: Any]).flatMap(makeHttpSyncOptions)
-        )
-    }
-
-    private func iosOptionsDictionary(_ options: IOSBackgroundLocationOptions) -> [String: Any] {
-        var dictionary: [String: Any] = [:]
-        dictionary["activityType"] = options.activityType?.stringValue
-        dictionary["pausesLocationUpdatesAutomatically"] = options.pausesLocationUpdatesAutomatically
-        dictionary["showsBackgroundLocationIndicator"] = options.showsBackgroundLocationIndicator
-        dictionary["useSignificantChanges"] = options.useSignificantChanges
-        dictionary["deferredUpdatesDistance"] = options.deferredUpdatesDistance
-        dictionary["deferredUpdatesInterval"] = options.deferredUpdatesInterval
-        return dictionary
-    }
-
-    private func makeIOSOptions(_ dictionary: [String: Any]) -> IOSBackgroundLocationOptions {
-        return IOSBackgroundLocationOptions(
-            activityType: (dictionary["activityType"] as? String).flatMap(IOSBackgroundActivityType.init(fromString:)),
-            pausesLocationUpdatesAutomatically: dictionary["pausesLocationUpdatesAutomatically"] as? Bool,
-            showsBackgroundLocationIndicator: dictionary["showsBackgroundLocationIndicator"] as? Bool,
-            useSignificantChanges: dictionary["useSignificantChanges"] as? Bool,
-            deferredUpdatesDistance: dictionary["deferredUpdatesDistance"] as? Double,
-            deferredUpdatesInterval: dictionary["deferredUpdatesInterval"] as? Double
-        )
-    }
-
-    private func activityOptionsDictionary(_ options: ActivityRecognitionOptions) -> [String: Any] {
-        var dictionary: [String: Any] = [:]
-        dictionary["enabled"] = options.enabled
-        dictionary["interval"] = options.interval
-        dictionary["stopOnStill"] = options.stopOnStill
-        dictionary["minimumConfidence"] = options.minimumConfidence
-        return dictionary
-    }
-
-    private func makeActivityOptions(_ dictionary: [String: Any]) -> ActivityRecognitionOptions {
-        return ActivityRecognitionOptions(
-            enabled: dictionary["enabled"] as? Bool,
-            interval: dictionary["interval"] as? Double,
-            stopOnStill: dictionary["stopOnStill"] as? Bool,
-            minimumConfidence: dictionary["minimumConfidence"] as? Double
-        )
-    }
-
-    private func httpSyncOptionsDictionary(_ options: BackgroundHttpSyncOptions) -> [String: Any] {
-        var dictionary: [String: Any] = [
-            "url": options.url
-        ]
-        dictionary["method"] = options.method?.stringValue
-        dictionary["headers"] = options.headers
-        dictionary["batch"] = options.batch
-        dictionary["batchSize"] = options.batchSize
-        dictionary["syncThreshold"] = options.syncThreshold
-        dictionary["syncInterval"] = options.syncInterval
-        dictionary["retry"] = options.retry
-        dictionary["maxRetries"] = options.maxRetries
-        dictionary["bodyTemplate"] = options.bodyTemplate.map(metadataDictionary)
-        dictionary["autoClear"] = options.autoClear
-        return dictionary
-    }
-
-    private func makeHttpSyncOptions(_ dictionary: [String: Any]) -> BackgroundHttpSyncOptions? {
-        guard let url = dictionary["url"] as? String else {
-            return nil
-        }
-        return BackgroundHttpSyncOptions(
-            url: url,
-            method: (dictionary["method"] as? String).flatMap(BackgroundHttpMethod.init(fromString:)),
-            headers: dictionary["headers"] as? [String: String],
-            batch: dictionary["batch"] as? Bool,
-            batchSize: dictionary["batchSize"] as? Double,
-            syncThreshold: dictionary["syncThreshold"] as? Double,
-            syncInterval: dictionary["syncInterval"] as? Double,
-            retry: dictionary["retry"] as? Bool,
-            maxRetries: dictionary["maxRetries"] as? Double,
-            bodyTemplate: (dictionary["bodyTemplate"] as? [String: Any]).map(makeMetadata),
-            autoClear: dictionary["autoClear"] as? Bool
-        )
-    }
-
-    private func backgroundLocation(_ location: StoredBackgroundLocation) -> BackgroundLocation {
-        return BackgroundLocation(
-            id: location.id,
-            source: location.source,
-            isFromBackground: location.isFromBackground,
-            provider: location.provider,
-            mocked: location.mocked,
-            recordedAt: location.recordedAt,
-            activity: location.activity,
-            battery: location.battery,
-            coords: location.coords,
-            timestamp: location.timestamp
-        )
-    }
-
-    private func nullableDouble(_ value: Any?) -> NullableDouble? {
-        guard let value = value as? Double else { return nil }
-        return .second(value)
-    }
-
-    private func variantDouble(_ value: NullableDouble?) -> Double? {
-        return value?.asType(Double.self)
-    }
-
-    private func metadataDictionary(
-        _ metadata: [String: Variant_NullType_Bool_String_Double]
-    ) -> [String: Any] {
-        var dictionary: [String: Any] = [:]
-        metadata.forEach { key, value in
-            switch value {
-            case .first:
-                dictionary[key] = ["__nitroNull": true]
-            case .second(let bool):
-                dictionary[key] = bool
-            case .third(let string):
-                dictionary[key] = string
-            case .fourth(let double):
-                dictionary[key] = double
-            }
-        }
-        return dictionary
-    }
-
-    private func bodyTemplateDictionary(
-        _ metadata: [String: Variant_NullType_Bool_String_Double]
-    ) -> [String: Any] {
-        var dictionary: [String: Any] = [:]
-        metadata.forEach { key, value in
-            switch value {
-            case .first:
-                dictionary[key] = NSNull()
-            case .second(let bool):
-                dictionary[key] = bool
-            case .third(let string):
-                dictionary[key] = string
-            case .fourth(let double):
-                dictionary[key] = double
-            }
-        }
-        return dictionary
-    }
-
-    private func makeMetadata(_ dictionary: [String: Any]) -> [String: Variant_NullType_Bool_String_Double] {
-        var metadata: [String: Variant_NullType_Bool_String_Double] = [:]
-        dictionary.forEach { key, value in
-            switch value {
-            case let nullMarker as [String: Any] where nullMarker["__nitroNull"] as? Bool == true:
-                metadata[key] = .first(NullType.null)
-            case let bool as Bool:
-                metadata[key] = .second(bool)
-            case let string as String:
-                metadata[key] = .third(string)
-            case let number as NSNumber:
-                metadata[key] = .fourth(number.doubleValue)
-            default:
-                metadata[key] = .third(String(describing: value))
-            }
-        }
-        return metadata
-    }
-
-    private func motionActivityType(_ activity: CMMotionActivity) -> DetectedActivityType {
-        if activity.automotive {
-            return .invehicle
-        }
-        if activity.cycling {
-            return .onbicycle
-        }
-        if activity.running {
-            return .running
-        }
-        if activity.walking {
-            return .walking
-        }
-        if activity.stationary {
-            return .still
-        }
-        return .unknown
-    }
-
-    private func motionConfidence(_ confidence: CMMotionActivityConfidence) -> Double {
-        switch confidence {
-        case .low:
-            return 25
-        case .medium:
-            return 60
-        case .high:
-            return 95
-        @unknown default:
-            return 0
-        }
-    }
-
-    private func mapActivityType(_ activityType: IOSBackgroundActivityType?) -> CLActivityType {
-        switch activityType {
-        case .automotivenavigation:
-            return .automotiveNavigation
-        case .fitness:
-            return .fitness
-        case .othernavigation:
-            return .otherNavigation
-        case .airborne:
-            if #available(iOS 12.0, *) {
-                return .airborne
-            }
-            return .other
-        case .other, nil:
-            return .other
-        @unknown default:
-            return .other
-        }
     }
 
     private func permissionResult() -> BackgroundPermissionResult {
